fix: terraform bugs + some pkg refactoring (#1235)
* fix: initialization of pipelines

* exp: add some logging

* feat: implement static vars

* add configs

* chore: add disable cleanup workspace option

* chore: fixes to show root directory output

* chore: more style guide notes

* chore: rename and migrate interceptors

* chore: refactor interceptors package

* fix: test with executing step

* chore: move temporal client package

* chore: update to new temporal pkg path

* chore: move temporalzap into `pkg/temporal/temporalzap`

* fix: exp using %w log.Fatal

* chore: update references to temporalzap

* chore: move shortid out of common

* more cleanup

* chore: change `shortid` package

* exp: fix fatal log

* chore: move uploader into `aws/s3uploader`

* chore: move `pkg/uploader` into `pkg/aws/s3uploader`

* chore: update references to `uploader`

+++ pkg/terraform/workspace/root.go
@@ -13,6 +13,10 @@ const (
 	defaultFileExecPermissions fs.FileMode = 0777
 )
 
+func (w *workspace) Root() string {
+	return w.root
+}
+
 // createRoot creates a new root directory for the workspace
 func (w *workspace) createRoot() error {
 	dir, err := os.MkdirTemp(w.tmpDirRoot, "workspace")
@@ -43,6 +47,10 @@ func (w *workspace) writeFile(path string, byts []byte, perms fs.FileMode) error
 
 // cleanup cleans up the root directory and all contents
 func (w *workspace) cleanup() error {
+	if w.DisableCleanup {
+		return nil
+	}
+
 	if err := os.RemoveAll(w.root); err != nil {
 		return fmt.Errorf("unable to remove %s: %w", w.root, err)
 	}
